+++ plugins/HtmlOutput/HtmlOutput.php
@@ -27,7 +27,7 @@ use XSLTProcessor;
  * Class HtmlOutput
  * @package IGCMS\Plugins
  */
-class HtmlOutput extends Plugin implements SplObserver, OutputStrategyInterface, ResourceInterface {
+class HtmlOutput extends Plugin implements SplObserver, OutputStrategyInterface {
   /**
    * @var string
    */
@@ -116,20 +116,45 @@ class HtmlOutput extends Plugin implements SplObserver, OutputStrategyInterface,
     }
     $this->cfg = self::getXML();
     $this->registerThemes($this->cfg);
-    $robots = $this->cfg->matchElement("robots", "domain", HTTP_HOST);
-    if (is_null($robots)) {
-      throw new Exception("Unable to match robots element to domain");
-    }
-    if (!$robots->hasAttribute("domain")) {
-      Logger::user_warning(_("Using default robots value (without domain match)"));
-    }
-    $this->metaRobots = $robots->getAttribute("meta");
-    if (stream_resolve_include_path(ROBOTS_TXT) && !@unlink(ROBOTS_TXT)) {
-      Logger::error(sprintf(_("Unable to delete %s file"), ROBOTS_TXT));
-    }
     if (is_null($this->favIcon)) {
       $this->favIcon = find_file($this->pluginDir."/".self::FAVICON);
     }
+    $cacheKey = apc_get_key(__FUNCTION__);
+    $useCache = false;
+    $cfgMtime = filemtime(find_file($this->pluginDir."/".$this->className.".xml"));
+    if (apc_exists($cacheKey)) {
+      $useCache = apc_is_valid_cache($cacheKey, $cfgMtime);
+    }
+    if ($useCache) {
+      return;
+    }
+    $metaRobots = "all";
+    $robotsFile = "User-agent: *";
+    $disallow = [];
+    $allowDomain = $this->cfg->matchElement("allow_domain", null, HTTP_HOST);
+    if (is_null($allowDomain)) {
+      $metaRobots = "noindex, nofollow";
+      $robotsFile = "User-agent: *\nDisallow: /";
+      $disallowDomain = $this->cfg->matchElement("disallow_domain", null, HTTP_HOST);
+      if (!is_null($disallowDomain)) {
+        Logger::user_warning(_("Robots disallowed due to no domain match"));
+      }
+    } else {
+      /** @var DOMElementPlus $disallow */
+      foreach ($this->cfg->getElementsByTagName('disallow') as $disallowElm) {
+        $disallow[] = $disallowElm->nodeValue;
+      }
+      if (count($disallow)) {
+        $robotsFile .= "\nDisallow: ".implode("\nDisallow: ", $disallow);
+      } else {
+        $robotsFile .= "\nDisallow:";
+      }
+    }
+    $this->metaRobots = $metaRobots;
+    if (!file_put_contents(ROBOTS_TXT, $robotsFile)) {
+      Logger::error(sprintf(_("Unable to save %s file"), ROBOTS_TXT));
+    }
+    apc_store_cache($cacheKey, $cfgMtime, __FUNCTION__);
   }
 
   /**
@@ -155,28 +180,6 @@ class HtmlOutput extends Plugin implements SplObserver, OutputStrategyInterface,
 
     // add root template files
     $this->addThemeFiles($cfg->documentElement);
-  }
-
-  /**
-   * @param string $filePath
-   * @return bool
-   */
-  public static function isSupportedRequest ($filePath) {
-    return $filePath === ROBOTS_TXT;
-  }
-
-  /**
-   * @return void
-   * @throws Exception
-   */
-  public static function handleRequest () {
-    $robots = self::getXML()->matchElement("robots", "domain", HTTP_HOST);
-    if (is_null($robots)) {
-      new ErrorPage("No matching robots element", 404);
-    }
-    header('Content-Type: text/plain');
-    echo $robots->nodeValue;
-    exit;
   }
 
   /**