+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - New plugin TreePresenter
  - Plugin Canonical
  - Open Graph protocol support
-
-## Added
  - Inputvar supports textarea set
+
+### Changed
+ - Improve robots syntax
 
 ## Changed
  - Hideable icon is defined by css